Allow fractional maximum points
change typecasts from int to float in
- updateSeparateTestQuestions
- updateFixedQuestions
- updateRandomQuestions
when calling $tst_run->setMaxPoints()

+++ components/ILIAS/Course/classes/Objectives/class.ilLOTestQuestionAdapter.php
@@ -457,7 +457,7 @@ class ilLOTestQuestionAdapter
                 $tst_run->addQuestion($qst_id);
                 $points += ilCourseObjectiveQuestion::_lookupMaximumPointsOfQuestion($qst_id);
             }
-            $tst_run->setMaxPoints((int) $points);
+            $tst_run->setMaxPoints((float) $points);
         }
     }
 
@@ -474,7 +474,7 @@ class ilLOTestQuestionAdapter
                 $tst_run->addQuestion($id);
                 $points += ilCourseObjectiveQuestion::_lookupMaximumPointsOfQuestion($id);
             }
-            $tst_run->setMaxPoints((int) $points);
+            $tst_run->setMaxPoints((float) $points);
         }
     }
 
@@ -502,7 +502,7 @@ class ilLOTestQuestionAdapter
                     $points += ilCourseObjectiveQuestion::_lookupMaximumPointsOfQuestion($qst);
                 }
             }
-            $tst_run->setMaxPoints((int) $points);
+            $tst_run->setMaxPoints((float) $points);
         }
     }
